fixup friend index route
