+++ app/routes/_app.friends._index/components/friend-menu.tsx
@@ -1,0 +1,12 @@
+import { Button } from "@/components";
+import { Link } from "@remix-run/react";
+
+export function FriendMenu() {
+	return (
+		<div>
+			<Button asChild variant={"outline"} size={"sm"}>
+				<Link to={"/explore/people"}>Find Friends</Link>
+			</Button>
+		</div>
+	);
+}
